refactor(ui): move coordinator-admin shell + lifecycle into coordinator-admin/lifecycle.ts
Finish the decomposition started across the coordinator-admin stack.
renderShell, initCoordinatorAdminTab, and loadCoordinatorAdminData now
live under coordinator-admin/lifecycle.ts alongside the factory wiring.
coordinator-admin.tsx becomes a barrel that re-exports the two public
entrypoints (initCoordinatorAdminTab, loadCoordinatorAdminData).
